Move MockClock into _core. No functional changes.
